requirements.txt
streamlit>=1.35.0
python-dotenv>=1.0.1
openai>=1.0.0
chromadb>=0.5.0
langchain-chroma>=0.1.1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
chainlit
rank-bm25
langchain-community
sentence-transformers
langchain-core
python-magic
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0